xfstests: fix a few build warnings
This patch fixes a few build warnings.  I have built the code using
i386, x86_64, and ia64 architectures and each has ends up with
complaints of one sort or anther.  This gets rid of all of them
*except* those reported by files under the "ltp" (Linux Test
Project) sub-tree.

+++ src/aio-dio-regress/aiodio_sparse2.c
@@ -227,7 +227,8 @@ void aiodio_sparse(char *filename, int align, int writesize, int startoffset, in
 		unsigned char *badbuf;
 
 		if (debug)
-			fprintf(stderr, "seek to %ld and read %d\n", offset, writesize);
+			fprintf(stderr, "seek to %lld and read %d\n",
+				(long long) offset, writesize);
 		lseek(fd, offset, SEEK_SET);
 		if (read(fd, bufptr, writesize) < writesize) {
 			fprintf(stderr, "short read() at offset %lld\n",
